pyproject.toml
[tool.poetry]
name = "nmt"
version = "0.1.0"
description = "NMT-based Translator"
authors = ["MSWon"]

[tool.poetry.scripts]
nmt = "nmt.cli:main"

[tool.poetry.dependencies]
python = "~3.7"

Flask = "2.1.0"
gunicorn = "20.1.0"
requests = "2.27.1"

torch = "1.11.0"
torchaudio = "0.11.0"
torchvision = "0.12.0"
tensorboard = "2.8.0"
sacrebleu = "2.0.0"
sentencepiece = "0.1.96"
numpy = "1.21.6"
tqdm = "4.64.0"
PyYAML = "6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
